show profile pics to the side of each chat
added Volley to gradle
and are now showing FB profile pics on side of each chat

rewrote list adapter for the ChatsActivity to use a Custom List Adapter
that has a ChatsMeta object and an imageURL.

next fix: make latest chatsMeta appear on top. right now the ordering
is inconsistent for some reason.