disabled refreshing chatsMeta in ChatActivity when not in ChatActivity
the Chat Rows in ChatActivity were refreshing upon any update to
ChatsMeta table, even when we weren’t in the class because we didn’t
detach listener.

not sure how to detach listener with all the complicated callbacks i
used, so I added a boolean that checks whether the user is in the
activity, this boolean activates in the onPause and onStart() methods.